models/wilor_hands/requirements.txt
numpy<2
opencv-python
pyrender

pytorch-lightning==2.0.9
torch==2.0.1
torchvision==0.15.2

scikit-image
smplx==0.1.28
yacs
chumpy @ git+https://github.com/mattloper/chumpy
timm
einops
xtcocotools
pandas
pyyaml
hydra-core
hydra-submitit-launcher
hydra-colorlog
pyrootutils
rich
webdataset
gradio
ultralytics==8.1.34
